TST: more work on using asyncio across threads correctly in tests
